ats-system/backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
pytesseract
Pillow
sentence-transformers
scikit-learn
spacy
SQLAlchemy
pydantic
